requirements.txt

absl-py==2.3.0
aioice==0.10.1
aiortc==1.13.0
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
av==14.4.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dnspython==2.7.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
google-pasta==0.2.0
grpcio==1.73.0
h5py==3.14.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
narwhals==1.42.1
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
protobuf==5.29.5
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.45.1
streamlit-webrtc==0.62.4
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
